Moved TimeVaryingValues to temporal package
