+++ examples/jobs/src/main/scala/com/example/jobs/Jobs.scala
@@ -3,7 +3,6 @@ package com.example.jobs
 import indigo._
 import indigoextras.jobs.Job
 import indigoextras.jobs.JobName
-import indigoextras.interleaved.IncreaseTo
 
 final case class ChopDown(index: Int, position: Point) extends Job {
   val isLocal: Boolean = false
