+++ examples/jobs/src/main/scala/com/example/jobs/Model.scala
@@ -3,7 +3,6 @@ package com.example.jobs
 import indigo._
 import indigo.shared.EqualTo._
 import indigoextras.jobs.JobMarketEvent
-import indigoextras.interleaved.IncreaseTo
 
 final case class Model(bob: Bob, grove: Grove, woodPiles: List[Wood], woodCollected: Int) {
